Module compiles, cairo off. M    datareduction/src/SectorList.cc M    datareduction/src/cairo-utils.cc M    datareduction/src/HoughTransformStraight.cc M    datareduction/src/SectorBasic.cc M    datareduction/src/PXDLadderList.cc M    datareduction/src/HoughTransformSine.cc M    datareduction/src/SectorArc.cc M    datareduction/src/SectorStraight.cc M    datareduction/src/DataReduction.cc M    datareduction/src/PXDLadder.cc M    datareduction/src/HoughTransformBasic.cc M    datareduction/SConscript M    datareduction/examples/Dummy.py
git-svn-id: https://belle2.cc.kek.jp/svn/trunk/software@1434 b38c6c8f-bee8-465d-8448-9abe6915d496

+++ tracking/modules/datareduction/include/cairo-utils.h
@@ -1,8 +1,7 @@
 #ifndef CAIRO_UTILS_H
 #define CAIRO_UTILS_H
-
-#include "cairo.h"
-#include "cairo-pdf.h"
+#include <cairo.h>
+#include <cairo-pdf.h>
 
 #include <vector>
 #include <string>
@@ -47,7 +46,7 @@ public:
   static cairo_color sample(std::vector<cairo_color> cmap, double value, double min, double max) {
     int steps = cmap.size() - 1;
     double position = value / fabs(max - min) * steps;
-    int left = floor(position);
+    int left = (int)floor(position);
     position -= left;
     if (max < min) {
       left = steps - left - 1;
